+++ examples/AutoBenchmark/generate_readme.py
@@ -1,0 +1,192 @@
+#!/usr/bin/python3
+#
+# Python script that regenerates the README.md from the embedded template. Uses
+# ./generate_table.awk to regenerate the ASCII tables from the various *.txt
+# files.
+
+from subprocess import check_output
+
+nano_results = check_output(
+    "./generate_table.awk < nano.txt", shell=True, text=True)
+micro_results = check_output(
+    "./generate_table.awk < micro.txt", shell=True, text=True)
+samd_results = check_output(
+    "./generate_table.awk < samd.txt", shell=True, text=True)
+esp8266_results = check_output(
+    "./generate_table.awk < esp8266.txt", shell=True, text=True)
+stm32_results = check_output(
+    "./generate_table.awk < stm32.txt", shell=True, text=True)
+esp32_results = check_output(
+    "./generate_table.awk < esp32.txt", shell=True, text=True)
+#teensy32_results = check_output(
+#    "./generate_table.awk < teensy32.txt", shell=True, text=True)
+teensy32_results = 'TBD'
+
+print(f"""\
+# AutoBenchmark
+
+This sketch measures the amount of time consumed by the `AceButton::check()`
+method when processing various button events. It uses a special
+`ProfilingButtonConfig` object that allows the program to inject button events
+into the library. The profiling numbers come from activating the `TimingStats`
+object that has been instrumented into the `AceButton::check()` method.
+
+Note that `ProfilingButtonConfig` class generates synthetic button events,
+bypassing the actual `digitalRead()` function. The `digitalRead()` function on
+an Arduino AVR platform (UNO, Nano, etc) is
+[known to be slow](https://forum.arduino.cc/index.php?topic=337578)
+which will add to the timing values shown below in actual usage.
+The [digitalWriteFast library](https://github.com/NicksonYap/digitalWriteFast)
+might be an alternative if speed is critical.
+
+**Version**: AceButton v1.8
+
+**DO NOT EDIT**: This file was auto-generated using `make README.md`. 
+
+## Dependencies
+
+This program depends on the following libraries:
+
+* [AceCommon](https://github.com/bxparks/AceCommon)
+* [AceButton](https://github.com/bxparks/AceButton)
+
+## How to Generate
+
+This requires the [AUniter](https://github.com/bxparks/AUniter) script
+to execute the Arduino IDE programmatically.
+
+The `Makefile` has rules to generate the `*.txt` results file for several
+microcontrollers that I usually support, but the `$ make benchmarks` command
+does not work very well because the USB port of the microcontroller is a
+dynamically changing parameter. I created a semi-automated way of collect the
+`*.txt` files:
+
+1. Connect the microcontroller to the serial port. I usually do this through a
+USB hub with individually controlled switch.
+2. Type `$ auniter ports` to determine its `/dev/ttyXXX` port number (e.g.
+`/dev/ttyUSB0` or `/dev/ttyACM0`).
+3. If the port is `USB0` or `ACM0`, type `$ make nano.txt`, etc.
+4. Switch off the old microontroller.
+5. Go to Step 1 and repeat for each microcontroller.
+
+The `generate_table.awk` program reads one of `*.txt` files and prints out an
+ASCII table that can be directly embedded into this README.md file. For example
+the following command produces the table in the Nano section below:
+
+```
+$ ./generate_table.awk < nano.txt
+```
+
+Fortunately, we no longer need to run `generate_table.awk` for each `*.txt`
+file. The process has been automated using the `generate_readme.py` script which
+will be invoked by the following command:
+```
+$ make README.md
+```
+
+The CPU times below are given in microseconds. The "samples" column is the
+number of `TimingStats::update()` calls that were made.
+
+## CPU Time Changes
+
+Version 1.8.1 adds benchmarks for STM32.
+
+## Arduino Nano
+
+* 16MHz ATmega328P
+* Arduino IDE 1.8.13
+* Arduino AVR Boards 1.8.3
+* `micros()` has a resolution of 4 microseconds
+
+```
+{nano_results}
+```
+
+## Sparkfun Pro Micro
+
+* 16 MHz ATmega32U4
+* Arduino IDE 1.8.13
+* SparkFun AVR Boards 1.1.13
+* `micros()` has a resolution of 4 microseconds
+
+```
+{micro_results}
+```
+
+## SAMD21 M0 Mini
+
+* 48 MHz ARM Cortex-M0+
+* Arduino IDE 1.8.13
+* Sparkfun SAMD Core 1.8.1
+
+```
+{samd_results}
+```
+
+## ESP8266
+
+* NodeMCU 1.0 clone, 80MHz ESP8266
+* Arduino IDE 1.8.13
+* ESP8266 Boards 2.7.4
+
+```
+{esp8266_results}
+```
+
+## STM32
+
+* STM32 "Blue Pill", STM32F103C8, 72 MHz ARM Cortex-M3
+* Arduino IDE 1.8.13
+* STM32duino 1.9.0
+
+```
+{stm32_results}
+```
+
+## ESP32
+
+* ESP32-01 Dev Board, 240 MHz Tensilica LX6
+* Arduino IDE 1.8.13
+* ESP32 Boards 1.0.4
+
+```
+{esp32_results}
+```
+
+## Teensy 3.2
+
+* 96 MHz ARM Cortex-M4
+* Arduino IDE 1.8.13
+* Teensyduino 1.53
+* Compiler options: "Faster"
+
+**Stale Data**: I am no longer able to program Teensy from my computer. Maybe it
+was the upgrade from Ubuntu 18.04 to 20.04. Maybe it was upgrading Arduino IDE
+1.8.9 to 1.8.13. Maybe it was upgrading Teensyduino. I don't know. I think the
+following data is from AceButton v1.5.
+
+```
+Sizes of Objects:
+sizeof(AceButton): 16
+sizeof(ButtonConfig): 24
+sizeof(Encoded4To2ButtonConfig): 28
+sizeof(Encoded8To3ButtonConfig): 28
+sizeof(EncodedButtonConfig): 36
+sizeof(LadderButtonConfig): 36
+
+CPU:
++------------------------+-------------+---------+
+|button event            | min/avg/max | samples |
+|------------------------+-------------+---------+
+|idle                    |   2/  3/  5 | 1988    |
+|press/release           |   2/  3/  6 | 1988    |
+|click                   |   2/  3/  6 | 1988    |
+|double click            |   2/  3/  6 | 1988    |
+|long press/repeat press |   2/  3/  6 | 1987    |
+|Encode4To2ButtonConfig  |   6/ 11/ 14 | 1972    |
+|Encode8To3ButtonConfig  |  16/ 27/ 30 | 1942    |
+|EncodeButtonConfig      |   8/ 20/ 22 | 1957    |
+|LadderButtonConfig      |  16/ 25/ 33 | 1945    |
++------------------------+-------------+---------+
+```
+""")